+++ cognite/neat/_session/_show.py
@@ -98,7 +98,7 @@ class ShowDataModelAPI(ShowBaseAPI):
         self.implements = ShowDataModelImplementsAPI(self._state)
 
     def __call__(self) -> Any:
-        if not self._state.rule_store.empty:
+        if self._state.rule_store.empty:
             raise NeatSessionError("No data model available. Try using [bold].read[/bold] to read a new data model.")
         last_target = self._state.rule_store.provenance[-1].target_entity
         rules = last_target.dms or last_target.information